Move `LangCode` to own module
